Refactored gpt and qwen samples into llm_infernce
