+++ modules/dnn/CMakeLists.txt
@@ -421,7 +421,7 @@ endif()
 ocv_option(WITH_ONNXRUNTIME_GENAI "Build with ONNX Runtime GenAI support" OFF IF WITH_ONNXRUNTIME)
 ocv_option(DOWNLOAD_ONNXRUNTIME_GENAI "Download ONNX Runtime GenAI prebuilt binaries" OFF IF WITH_ONNXRUNTIME_GENAI)
 
-set(ONNXRUNTIME_GENAI_VERSION "0.12.1" CACHE STRING "ONNX Runtime GenAI version to download (prebuilt binaries)")
+set(ONNXRUNTIME_GENAI_VERSION "0.13.1" CACHE STRING "ONNX Runtime GenAI version to download (prebuilt binaries)")
 
 if(WITH_ONNXRUNTIME_GENAI AND WITH_ONNXRUNTIME)
   include("${OpenCV_SOURCE_DIR}/cmake/FindONNX.cmake")
@@ -443,20 +443,20 @@ if(WITH_ONNXRUNTIME_GENAI AND WITH_ONNXRUNTIME)
     if(WIN32)
       if(X86_64 OR CMAKE_SIZEOF_VOID_P EQUAL 8)
         set(_ortga_filename "onnxruntime-genai-${ONNXRUNTIME_GENAI_VERSION}-win-x64.zip")
-        set(_ortga_md5 "5fd68ddea370e12a391605bf767c2fce")
+        set(_ortga_md5 "f97d620d3b75a76ac2735e5241c6e924")
       elseif(ARM64 OR AARCH64)
         set(_ortga_filename "onnxruntime-genai-${ONNXRUNTIME_GENAI_VERSION}-win-arm64.zip")
-        set(_ortga_md5 "5e6fdfcea9e3668d0f5ef9c6fc071a41")
+        set(_ortga_md5 "fee7a457d3c9a7a5c5313fbe70e424cf")
       endif()
     elseif(APPLE)
       if(ARM64 OR AARCH64 OR CMAKE_OSX_ARCHITECTURES MATCHES "arm64")
         set(_ortga_filename "onnxruntime-genai-${ONNXRUNTIME_GENAI_VERSION}-osx-arm64.tar.gz")
-        set(_ortga_md5 "d817ca0181bd617e9c1d38e6221b3fdc")
+        set(_ortga_md5 "14d8af363b159b5267e238eeb98b7f1b")
       endif()
     elseif(UNIX)
       if(X86_64)
         set(_ortga_filename "onnxruntime-genai-${ONNXRUNTIME_GENAI_VERSION}-linux-x64.tar.gz")
-        set(_ortga_md5 "91f59d2f37849729c6c45d15942ec6c6")
+        set(_ortga_md5 "ad7f7723d102e08d1fbd4c6c5be40a5c")
       endif()
     endif()
